Ai_video_generator/requirements.txt
openai
python-dotenv
google-api-python-client
google-cloud
google-cloud-vision
google-generativeai
google-genai #The correct google install
flask
flask-cors
pymongo